+++ src/Commands/Masternode/PauseCommand.php
@@ -2,7 +2,9 @@
 
 namespace pxgamer\ArionumCLI\Commands\Masternode;
 
-use GuzzleHttp\Exception\GuzzleException;
+use pxgamer\Arionum\ApiException;
+use pxgamer\Arionum\Transaction;
+use pxgamer\ArionumCLI\ArionumException;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -25,18 +27,26 @@ final class PauseCommand extends MasternodeCommand
      * @param OutputInterface $output
      * @return void
      * @throws \Exception
-     * @throws GuzzleException
      */
     protected function execute(InputInterface $input, OutputInterface $output): void
     {
         parent::execute($input, $output);
 
         try {
-            $result = $this->sendCommand(self::COMMAND_VERSION_PAUSE);
+            $date = time();
+            $signature = $this->returnCommandSignature(Transaction::VERSION_MASTERNODE_PAUSE, $date);
 
-            $output->writeln('<info>Masternode pause command sent!</info>');
-            $output->writeln('<info>ID: '.$result['data'].'</info>');
-        } catch (\Exception $exception) {
+            $transaction = Transaction::makeMasternodePauseInstance($this->wallet->getAddress());
+
+            $transaction->setSignature($signature);
+            $transaction->setPublicKey($this->wallet->getPublicKey());
+            $transaction->setDate($date);
+
+            $transactionId = $this->arionumClient->sendTransaction($transaction);
+
+            $output->writeln('<info>Masternode `pause` command sent!</info>');
+            $output->writeln('<comment>Transaction id:</comment> '.$transactionId);
+        } catch (ApiException | ArionumException $exception) {
             $output->writeln('<error>'.$exception->getMessage().'</error>');
         }
     }